refactored AnnotationFeatureForm.class in its own file (attempt 2)
